==> 10_create_summary_and_questions.ipynb <==
import importlib

import src.valid_index
importlib.reload(src.valid_index)
from src.valid_index import get_popia_act_index

import src.file_tools
importlib.reload(src.file_tools)
from src.file_tools import read_processed_regs_into_dataframe, get_regulation_detail

import src.embeddings
importlib.reload(src.embeddings)
from src.embeddings import get_ada_embedding, num_tokens_from_string

import src.tree_tools
importlib.reload(src.tree_tools)
from src.tree_tools import build_tree_for_regulation, split_tree


file_list = []
file_list.append('./txt/popia_act.txt')
non_text_labels = ['Definition']

index_popia = get_popia_act_index()
df_popia, non_text = read_processed_regs_into_dataframe(file_list=file_list, valid_index_checker=index_popia, non_text_labels=non_text_labels)
tree_popia = build_tree_for_regulation("POPIA", df_popia, valid_index_checker=index_popia)

section_summary_with_embeddings = "./tmp/summary_popia_with_embedding.parquet"
section_questions_with_embeddings = "./tmp/summary_popia_questions_with_embedding.parquet"
headings_index_file = "./tmp/headings.csv"




import pandas as pd
import os

sectioned_df = pd.DataFrame([],columns = ["section", "text", "token_count"])
save_sectioned_df_to_file = "./tmp/popia_manual.csv"
exists = os.path.exists(save_sectioned_df_to_file)

if not exists:
    print("Loading the initial split of the tree. You will need to make changes to this as you see the data")
    # Starting at an particular parent node (can be the tree root or any child), this method splits up the 
    # branch into sections where the text does not exceed a certain word_count cap.
    sectioned_df = split_tree(tree_popia.root, df_popia, 750, index_popia)
else:
    print("Loading the currency split of the tree so you can continue generating summaries and questions")
    sectioned_df = pd.read_csv(save_sectioned_df_to_file, encoding="utf-8", sep="|")

print(f'Total number of sections: {len(sectioned_df)}')


Create or load the DataFrames that will hold the text index. Later we will add the embeddings to the same DataFrames. When we do so, loading the embeddings is slow from certain file formats like csv so we just start with a fast loading file format - parquet 

df_summary = None
if os.path.exists(section_summary_with_embeddings):
    df_summary = pd.read_parquet(section_summary_with_embeddings, engine='pyarrow')
    print(f"Summary data contains {len(df_summary)} lines of text")
    missing = len(df_summary[df_summary["text"] == ""])
    if missing > 0:
        print(f" -- of which there are {missing} lines that do not contain index text (e.g. sections with only definitions or indexes)")
else:
    print("Creating a new summary DataFrame")
    df_summary = pd.DataFrame([], columns = ["text", "section"])


df_questions = None
if os.path.exists(section_questions_with_embeddings):
    df_questions = pd.read_parquet(section_questions_with_embeddings, engine='pyarrow')
    print(f"Questions data contains {len(df_questions)} lines of text")    
    missing = len(df_questions[df_questions["text"] == ""])
    if missing > 0:
        print(f" -- of which there are {missing} lines that do not contain index text (e.g. sections with only definitions or indexes)")
else:
    print("Creating a new questions DataFrame")
    df_questions = pd.DataFrame([], columns = ["text", "section"])

index = None
if len(df_summary) != len(df_questions):
    print("The summary and the questions DataFrames do not have the same length")
else:
    index = len(df_summary)
    p = (index / len(sectioned_df)) * 100
    print(f'There are a total number of {len(sectioned_df)} sections to index')
    print(f"You have created {p:.2f} percent of your text index")


index = 120

## Create the text index 

This is a manual process. We call OpenAI and print out the answers in a format that is used to update the index text but this does need to be edited before it is added to the index so this is not automated.


import src.summarise_and_question
importlib.reload(src.summarise_and_question)
from src.summarise_and_question import get_summary_and_questions_for


#model = "gpt-3.5-turbo"
#model="gpt-4"
model="gpt-4-1106-preview"

reg_text = sectioned_df.loc[index]['text']
print("##############")
print(reg_text)
print("##############")

model_summary, model_questions = get_summary_and_questions_for(reg_text, model = model)

#format output
section = sectioned_df.loc[index]['section']
print(f'df_summary.loc[index, "section"] = "{section}"')
print(f'df_summary.loc[index, "text"] = "{model_summary}"')
print()
print(f'df_questions.loc[index, "section"] = "{section}"')
print(f'df_questions.loc[index, "text"] = "{model_questions}"')



df_summary.loc[index-1, "section"] = "114."

# Once the last four lines are manually checked, the edited result is copied here and the summary and question index is updated

df_summary.loc[index, "section"] = "115."
df_summary.loc[index, "text"] = "The name of the Act is the Protection of Personal Information Act, 2013, and it will start to apply on a date that the President announces in the Government Gazette. It's possible that various sections of the Act or different types of information and organisations may have different start dates."

df_questions.loc[index, "section"] = "115."
df_questions.loc[index, "text"] = "What is the formal name of the legislation governing personal information protection in South Africa, and when will it commence? | Are there different commencement dates for various sections of the personal information protection law or for different entities?"





index = index + 1
if index == len(sectioned_df):
    print("All done!")
else:
    next_section = sectioned_df.iloc[index]["section"]
    assert len(sectioned_df[sectioned_df["section"] == next_section]) == 1, "Huston, we have a problem"
    print(f'Next section is {next_section} which is on line {index}')
    p = ((index-1) / len(sectioned_df)) * 100
    print(f"You have completed {p:.2f} percent of your work")
    reg_text = sectioned_df.loc[index]['text']
    print("Next section")
    print("##############")
    print(reg_text)
    print("##############")



# Sometimes there are errors in the previous code block. We need to be careful when saving over any work we have already done so the 
# save step is a manual one which needs to be run regularly but without overwriting good data with bad data
df_summary.to_parquet(section_summary_with_embeddings, engine='pyarrow')
df_questions.to_parquet(section_questions_with_embeddings, engine='pyarrow')


#df_summary.drop(49, inplace=True)
df_summary[134:145]

#df_questions[40:55]

## Changing how the sections are chunked

From time to time we will see instances whe the initial chunk size needs to be adjusted and nodes need to be expanded or collapsed. We do this in two stages. First we can experiment with a new "token_limit_per_chunk" for the node in question and once we find the chunking solution we are after, we remove the old chunks and replace them with the new chunks.


node_str = "C.1(D)(v)"
# get the list of indicies that start with this string
index_list = sectioned_df.index[sectioned_df['section'].str.startswith(node_str)].tolist()
is_consecutive = all(x+1 == y for x, y in zip(index_list[:-1], index_list[1:]))
assert is_consecutive, "The list of indicies that start with the node string is not consecutive so the rest of the logic here will not hold"
start_index_to_replace = index_list[0]
end_index_to_replace = index_list[-1] + 1
print(f"This will remove {len(index_list)} row(s) from sectioned_df. From {start_index_to_replace} to {end_index_to_replace}")

# Get the new set of indicies assuming a different chunking length
node = tree_popia.get_node(node_str)
token_limit_per_chunk = 1300

tmp_df = split_tree(node, df_popia, token_limit_per_chunk, index_popia)

print(f"... and will replace them with {len(tmp_df)} row(s)")
print(tmp_df)


Replace the node and all its children with the new DataFrame with a different word_count limit

def replace_rows(original_df, updated_section_df, start_row, end_row):
    before = original_df.iloc[:start_row]
    after = original_df.iloc[end_row:]
    new_df = pd.concat([before, updated_section_df, after]).reset_index(drop=True)
    return new_df

print(f"The original data consisted of {sectioned_df} chunks")
sectioned_df = replace_rows(sectioned_df, tmp_df, start_row=start_index_to_replace, end_row=end_index_to_replace)
print(f"Post the update, the data consists of {sectioned_df} chunks")

Jesus saves! 
But only if he is happy with the results. Check first!

sectioned_df

sectioned_df.to_csv(save_sectioned_df_to_file, encoding="utf-8", sep="|", index = False)

## Definitions

import re
import pandas as pd

import src.embeddings
importlib.reload(src.embeddings)
from src.embeddings import get_ada_embedding, num_tokens_from_string

def count_leading_spaces(s):
    match = re.match(r'^\s*', s)  # Matches leading whitespace
    return len(match.group(0))

add_embeddings = False
definitions_to_process = '#Definition 1'
if definitions_to_process == '#Definition 1':
    popia_definitions_and_embeddings_file = "./tmp/popia_definitions_with_embeddings.csv"
    exclude_first_line = True
    start_line = 2
else:
    raise NotImplemented("Only implemented for popia Definition 1")

popia_manual_definitions = []
#raw_list = non_text['Definition']['#Definition 1']
raw_list = non_text['Definition'][definitions_to_process]
number_of_spaces = count_leading_spaces(raw_list[start_line])
if number_of_spaces % 4 != 0:
    raise ValueError(f"This line does not have an indent which is a multiple of 4: {raw_list[start_line]}")

current_line = raw_list[start_line]

current_line_number_of_spaces = count_leading_spaces(current_line)
if current_line_number_of_spaces != number_of_spaces:
    print(f"current_line_number_of_spaces: {current_line_number_of_spaces}")
    print(f'number_of_spaces: {number_of_spaces}')
    raise ValueError(f"This line does not have the correct indentation: {current_line}")

processing_table = False
for line_number in range(start_line,len(raw_list)-1):
    next_line = raw_list[line_number + 1]
    next_line_number_of_spaces = count_leading_spaces(next_line)
    if next_line_number_of_spaces % 4 != 0:
        raise ValueError(f"This line does not have an indent which is a multiple of 4: {next_line_number_of_spaces}")

    if current_line_number_of_spaces == next_line_number_of_spaces:
        current_line = current_line.lstrip()
        if "|" in current_line: # processing something with table formatting
            processing_table = True
            split_line = [x.strip() for x in current_line.split("|")]
            popia_manual_definitions.append(split_line)
        else:
            popia_manual_definitions.append(current_line)
        current_line = next_line
    else:
        current_line = current_line + "\n" + next_line

# add the last entry
current_line = current_line.lstrip()
if processing_table:
    split_line = [x.strip() for x in current_line.split("|")]
    popia_manual_definitions.append(split_line)
else:
    popia_manual_definitions.append(current_line.lstrip())

if processing_table:
    headings = popia_manual_definitions[0]
    popia_manual_definitions.pop(0)
    df_popia_definitions = pd.DataFrame(popia_manual_definitions, columns=headings) 
else:
    df_popia_definitions = pd.DataFrame(popia_manual_definitions, columns=["Definition"])


if add_embeddings:
    df_popia_definitions["Embedding"] =df_popia_definitions["Definition"].apply(get_ada_embedding)

df_popia_definitions["source"] = 'all'

#df_popia_definitions.to_csv(popia_definitions_and_embeddings_file, encoding = "utf-8", sep = "|", index = False)
df_popia_definitions.to_parquet("./inputs/popia_definitions.parquet", engine='pyarrow')


df_popia_definitions = pd.read_parquet("./inputs/popia_definitions.parquet", engine='pyarrow')

df_popia_definitions['source'] = 'all'

df_popia_definitions.to_parquet("./inputs/popia_definitions.parquet", engine='pyarrow')

## Section headings are also a good index

# step 1 is to remove all text that is not tagged as a heading leaving only the index and headings
toc_file = "./tmp/section_numbers_and_headings.txt" # Note this is a temporary file and will be deleted in a few cells time
df = df_popia
index = index_popia


written_references = set() # only write each reference once
with open(toc_file, 'w', encoding = 'utf-8') as f:
    for _, row in df.iterrows():
        if row['full_reference'] not in written_references:
            written_references.add(row['full_reference'])
            s = ' ' * row['Indent'] * 4 + row['Reference']
            # skip the text for into and legal
            if row['Heading'] and (row['Indent'] == 0 and row['Text'].strip() in index.exclusion_list):
                s = s
            elif row['Heading']:               
                s += " " + row['Text']
            f.write(s + '\n')

# step 2) Remove all lines that do not have text. Since the only text is for the headers, this removes everything that is not a header.
#         Note however that the remaining "headers" will not contain the (#headers) markdown so will be treated as text
import src.file_tools
importlib.reload(src.file_tools)
from src.file_tools import process_regulations

excon_headers = './tmp/non_empty_headings_excon.txt'  # Note this is a temporary file and will be deleted in a few cells time
files_as_list = []
files_as_list.append(toc_file)
df_toc, non_text_toc = process_regulations(files_as_list, valid_index_checker=index_popia, non_text_labels=non_text_labels)

for index, row in df_toc.iterrows():
    if row['Text'] != "":
        df_toc.at[index, 'Heading'] = True

tree_toc = build_tree_for_regulation("popia_toc", df_toc, valid_index_checker=index_popia)
l = tree_toc._list_node_children(tree_toc.root)
with open(excon_headers, 'w', encoding = 'utf-8') as f:
    f.write(l)

# step 3) The file that contains the headers (now as text because they are missing the (#heading) markdown) and load it up
#         as if it were the regs themselves
files_as_list = []
files_as_list.append(excon_headers)
df_non_empty_toc, non_text_toc = process_regulations(files_as_list, valid_index_checker=index_popia, non_text_labels=non_text_labels)
for index, row in df_non_empty_toc.iterrows():
    if row['Text'] != "":
        df_non_empty_toc.at[index, 'Heading'] = True

non_empty_tree_toc = build_tree_for_regulation("Excon", df_non_empty_toc, valid_index_checker=index_popia)


# Construct the full reference of each heading and use these as a key for a dictionary where the heading is the value
import os

def get_leaf_headings(root):
    leaf_headings = {}

    def recurse(node, heading):
        # Add ". " only if it's not the root node and the node's heading_text is not empty
        new_heading = heading + (". " + node.heading_text if heading and node.heading_text else node.heading_text)
        if not node.children:  # This is a leaf node.
            leaf_headings[node.full_node_name] = new_heading
        else:  # This is not a leaf node. We continue the recursion.
            for child in node.children:
                recurse(child, new_heading)

    recurse(root, '')  # We start the recursion from the root, with an empty heading.
    return leaf_headings

leaf_headings = get_leaf_headings(non_empty_tree_toc.root)
df_section_headings = pd.DataFrame(list(leaf_headings.items()), columns=['section', 'text'])
print(f'Popia Act contains {len(leaf_headings)} section headings')




df_section_headings.to_csv(headings_index_file, encoding = "utf-8", sep = "|", index = False)
if os.path.exists(toc_file):
    os.remove(toc_file)
if os.path.exists(excon_headers):
    os.remove(excon_headers)



df_index.to_parquet("./inputs/popia_index.parquet", engine='pyarrow')
# NOTE: At some point the heading sections were given an additional "0" at the end of the section which causes issues.
#        Confirm all the headings are "2." and not "2.0"

# Index

df_index = pd.read_parquet(section_questions_with_embeddings, engine='pyarrow')
df_index = df_index[df_index["text"] != ""] # remove rows that have 'text' == ""
# the 'text' column for the questions may contain multiple questions separated by a "|". The next line expands these rows
# so the value in 'text' only contains one question
df_index = df_index.drop("text", axis=1).join(df_index["text"].str.split("|", expand=True).stack().reset_index(level=1, drop=True).rename("text"))
df_index.reset_index(drop=True, inplace=True)
df_index["source"] = "question"

df_tmp = pd.read_parquet(section_summary_with_embeddings, engine='pyarrow')
df_tmp["source"] = "summary"

df_tmp_2 = pd.read_csv(headings_index_file, encoding = "utf-8", sep = "|")
df_tmp_2["source"] = "heading"

df_index = pd.concat([df_index, df_tmp, df_tmp_2], ignore_index = True)
df_index = df_index[df_index["text"]!= ""]
df_index = df_index[df_index["text"].notna()] # Remove any NaN's
df_index.reset_index(drop=True, inplace=True)

#df_index['Embedding'] = df_index['text'].apply(get_ada_embedding)


df_index['section'] = df_index['section'].astype(str)
df_index['text'] = df_index['text'].astype(str)
df_index['source'] = df_index['source'].astype(str)

df_index['Embedding'] = df_index['text'].apply(get_ada_embedding)

df_index.dtypes

df_index.to_parquet("./inputs/popia_index.parquet", engine='pyarrow')


df_tmp = pd.read_parquet("./inputs/popia_index.parquet", engine='pyarrow')

df_tmp

# if there is an error somewhere in the generation of the embedding and you need to find it, this is a hacky way to do that
increment = 10
for i in range(0, len(df_index), increment):
    chunk = df_index.iloc[i:i+increment].copy()
    chunk["Embedding"] = chunk["text"].apply(get_ada_embedding)
    df_index.loc[chunk.index, "Embedding"] = chunk["Embedding"]
    print(f"Completed {i+increment} lines")


#df_index
