==> 50_chatbot.ipynb <==
import importlib
import src.chat_bot
importlib.reload(src.chat_bot)
from src.chat_bot import PopiaAct


logging_level = 15
#logging_level = 20
log_file = ''

path_to_manual_as_csv_file = "./inputs/popia_manual.csv"
path_to_definitions_as_parquet_file = "./inputs/popia_definitions.parquet"
path_to_index_as_parquet_file = "./inputs/popia_index.parquet"


popia = PopiaAct(path_to_manual_as_csv_file, path_to_definitions_as_parquet_file, path_to_index_as_parquet_file, log_file=log_file, logging_level=logging_level)


print(popia.get_regulation_detail("56."))

#model_to_use="gpt-3.5-turbo"
#model_to_use = "gpt-4-1106-preview"
model_to_use="gpt-4"
#model_to_use="gpt-3.5-turbo-16k"

#user_context = "What is personal information?"
#user_context = "How should I protect personal information?"
#user_context = "What are the risks associated with storing personal information?"
#user_context = "Does POPIA apply to all industries?"
#user_context = "What does a data protection framework need to address?"



user_context = "As a medium size company, what do I need to do to comply with POPIA?"
user_context = "Do I need to employ additional staff?"
user_context = "What is an information officer?"
user_context = "Does every company need an information officer?"
user_context = "what do I need to do to comply with the act?"


popia.reset_conversation_history()
popia.user_provides_input(user_context, threshold=0.15, model_to_use=model_to_use, temperature=0, max_tokens=300)
print(popia.messages[-1]["content"])


popia.messages

import pandas as pd 
tmp_df = pd.read_parquet(path_to_index_as_parquet_file)


tmp_df.to_parquet(path_to_index_as_parquet_file, engine='pyarrow')

import src.embeddings
importlib.reload(src.embeddings)
from src.embeddings import get_ada_embedding, num_tokens_from_string


# Example values for the new row
question = 'What do I need to do to comply with POPIA?'
embedding = get_ada_embedding(question)

new_row = {
    'section': '56.',
    'text': question,
    'source': 'manual',
    'Embedding': embedding
}

# Adding the new row to the dataframe
tmp_df.loc[len(tmp_df)] = new_row

#(tmp_df[tmp_df["source"]=="question"])
#tmp_df.loc[486, "source"] = "manual"
tmp_df

tmp_df['section'] = tmp_df['section'].str.replace(r'\.0$', '.', regex=True)

tmp_df[tmp_df["section"].str.endswith(".")]

len(tmp_df)

import importlib
import src.valid_index
importlib.reload(src.valid_index)
from src.valid_index import get_popia_manual_index
import re

index = get_popia_manual_index()


s = " B.4(B)(ii)(a) of the Manual states that foreign currency may be made available within the single discretionary allowance limit of R1 million per calendar year to eligible parties. However, the specific limits for adults are not mentioned in the provided sections. Please refer to the relevant section(s) for more information on the limits for travel allowances for adults."
#s = "B.4(B)(ii)(a)"

print(index.extract_valid_reference(s))


patterns = [
        r'[A-Z]\.\d{0,2}',
        r'\([A-Z]\)',
        r'\(\b(i|ii|iii|iv|v|vi|vii|viii|ix|x|xi|xii|xiii|xiv|xv|xvi|xvii|xviii|xix|xx|xxi|xxii|xxiii|xxiv|xxv|xxvi|xxvii)\b\)',
        r'\([a-z]\)',
        r'\([a-z]{2}\)',
        r'\((?:[1-9]|[1-9][0-9])\)'
    ]
matches = [re.findall(pattern, s) for pattern in patterns]
matches

match = re.search(r'\([a-z]\)', s)
match

s[match.end()]

popia.get_regulation_detail("B.5(B)(iii)")

print(popia._create_system_message())

import pandas as pd
df_text_all = pd.read_parquet(path_to_index_as_parquet_file, engine='pyarrow')
df_text_all

popia.get_regulation_detail('I.(1)(E)(ii)')

#user_context = "What is a CFC Account?"
#user_context = "I am submitting documentation to the financial surveillance department. What information do I need to include in my submission?" # (A.3 (B))
#user_context = "Can a bank provide foreign currency denominated facilities to a South African Corporate?" #  (A.3 (B))
#user_context = "For how long does the approval for a facility remain valid?" #  (A.3 (B))
#user_context = "Other than authorised transaction requests, are there requirements for authorised dealers to submit information to the financial surveillance department?" #  (A.3 (B))
#user_context = "I have been asked to act on an approval issued to another bank. Can I do that?" #  (A.3 (B))
#####user_context = "--- Same as above but use "Authorised Dealer" instead of "Bank"
#user_context = "If I am concerned that as client may in breach of regulations, what should I do?" #  (A.3 (B))
#user_context = "What constitutes 'documentary evidence' for a foreign currency transaction?" #  (A.3 (C))
#user_context = "Is Swaziland part of the CMA?" #  (A.3 (E))
#user_context = "A client wants to buy Namibian Dollars. What approvals do they need before they can do this?" #  (A.3 (E))
#user_context = "A person from Lesotho wants to buy dollars from me. Can I sell them the dollars?" #  (A.3 (E))
#user_context = "A person from Namibia wants to buy rand from me. Can I sell them the rand?" #  (A.3 (E))
#user_context = "A treasury management company wants to trade dollars with me. What do I need to check before I can do this?" #  (A.4 (B))
#user_context = "A client wants to import goods and pay in foreign currency. What documentation do I need before I can assist with the transaction" #  (B.1 (B))
#user_context = "Over and above the price of the goods, what additional costs can be covered in foreign currency?" #  (B.1 (C))
#user_context = "Can advance payment be made in a foreign currency to cover imported goods?" #  (B.1 (D)(i))

#user_context = "Can advance payment be made in a foreign currency to cover imported capital goods?" #  (B.1 (D)(ii))
#user_context = "When paying for imports in cash, what documentation is required before providing the foreign currency?" #  (B.1 (D)(iii))

#user_context = "Can imported goods be financed over more than 12 months?" #  (B.1 (D)(iv))
#user_context = "How much money can an individual invest internationally?" #  ?
#user_context = "How much money can an individual take offshore?" #  ?

#user_context = "A media company want to list offshore. Are there specific conditions for this?"
# user_context = "How many rands can a resident take offshore in cash?"
#user_context = "Can I grant a foreign company a Rand facility to pay for goods they wish to purchase from a South African exporter?"
#user_context = "Can I support a non-residential entity with short-term rand trade finance for their exports from South Africa?"
#user_context = "Hi"
#user_context = "Who can trade gold?"
#user_context = "are there specific exchange control rules for religious organizations?"
#user_context = "How much foreign currency can a charity take offshore each year?"
